api-service/requirements.txt
requests
fastapi
uvicorn
numpy
pillow
python-multipart
aiofiles
httpx
pydantic
pandas
torch
torchvision
grpcio
tritonclient
tritonclient[grpc]
tritonclient[utils]